Fixes default font loader plugin and puts it behind feature flag
