+++ src/text/mod.rs
@@ -1,12 +1,16 @@
 //! Components and text logic.
 
 mod context;
-mod default_font;
 mod font;
 mod font_loader;
 mod vello_text;
 
-pub(crate) use default_font::DefaultFontPlugin;
 pub use font::VelloFont;
 pub(crate) use font_loader::VelloFontLoader;
 pub use vello_text::{VelloTextAnchor, VelloTextSection, VelloTextStyle};
+
+#[cfg(feature = "default_font")]
+mod default_font_plugin;
+
+#[cfg(feature = "default_font")]
+pub(crate) use default_font_plugin::DefaultFontPlugin;
